🚚 Rename:  Transitions 폴더 내 Collapse, Slide, Fade, Scale 폴더 생성해 따로 분리 & 각각 mdx 작성
