+++ src/components/general/Transitions/Collapse/Collapse.mdx
@@ -1,0 +1,41 @@
+import { Meta, Canvas, ArgTypes, Stories } from '@storybook/blocks';
+import * as CollapseStories from './Collapse.stories';
+
+<Meta of={CollapseStories} />
+
+# Collapse
+
+Expand(확장)/Collapse(축소) Transition<br />
+(Transition 컴포넌트를 기반으로 만들어짐)
+
+## Import
+
+```js
+import { Collapse } from '@/components/general/Transitions';
+```
+
+## Usage
+
+### Default Collapse Transition
+
+- children의 start edge에서부터 expand 되고 end edge에서부터 collapse 됨
+
+<Canvas of={CollapseStories.DefaultCollapseTransition} />
+
+### Orientation
+
+- `orientation`: transition 방향
+
+<Canvas of={CollapseStories.Orientation} />
+
+### Collapsed Size
+
+- `collapsedSize`: collapse될 때 container의 크기
+  - orientation='vertical'인 경우, container의 height
+  - orientation='horizontal'인 경우, container의 width
+
+<Canvas of={CollapseStories.CollapsedSize} />
+
+## Props
+
+<ArgTypes of={CollapseStories} sort="alpha" />
