+++ src/components/general/Transitions/Collapse/Collapse.stories.tsx
@@ -4,7 +4,7 @@ import Collapse from './Collapse';
 import { Stack } from '@/components/layout/Stack';
 import { Box } from '@/components/layout/Box';
 import { Button } from '@/components/general/Button';
-import { TRANSITION_ARG_TYPES } from './Transition.constants';
+import { TRANSITION_ARG_TYPES } from '../Transition.constants';
 
 const meta: Meta<typeof Collapse> = {
   component: Collapse,
