+++ src/components/general/Transitions/Collapse/index.ts
@@ -1,0 +1,1 @@
+export { default as Collapse } from './Collapse';
