+++ src/components/general/Transitions/Fade/Fade.stories.tsx
@@ -1,20 +1,20 @@
 import { useState } from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import Scale from './Scale';
+import Fade from './Fade';
 import { Stack } from '@/components/layout/Stack';
 import { Box } from '@/components/layout/Box';
 import { Button } from '@/components/general/Button';
-import { TRANSITION_ARG_TYPES } from './Transition.constants';
+import { TRANSITION_ARG_TYPES } from '../Transition.constants';
 
-const meta: Meta<typeof Scale> = {
-  component: Scale,
+const meta: Meta<typeof Fade> = {
+  component: Fade,
   argTypes: TRANSITION_ARG_TYPES
 };
 
 export default meta;
-type Story = StoryObj<typeof Scale>;
+type Story = StoryObj<typeof Fade>;
 
-const ScaleTemplate = ({ ...props }) => {
+const FadeTemplate = ({ ...props }) => {
   const [transitionIn, setTransitionIn] = useState(false);
 
   const toggleTransition = () => {
@@ -26,32 +26,24 @@ const ScaleTemplate = ({ ...props }) => {
       <Button onClick={toggleTransition} style={{ maxWidth: 'max-content' }}>
         {transitionIn ? 'Disappear' : 'Appear'}
       </Button>
-      <Scale in={transitionIn} {...props}>
+      <Fade in={transitionIn} {...props}>
         <Box
           style={{ backgroundColor: 'yellow-300', width: 100, height: 100 }}
         />
-      </Scale>
+      </Fade>
     </Stack>
   );
 };
 
-export const DefaultScaleTransition: Story = {
+export const DefaultFadeTransition: Story = {
   render: (args) => (
     <Stack direction="row" spacing={20}>
-      <ScaleTemplate {...args} />
-      <ScaleTemplate
-        easing={{
-          enter: 'emphasized-decelerate',
-          exit: 'emphasized-accelerate'
-        }}
+      <FadeTemplate {...args} />
+      <FadeTemplate
+        easing="standard"
+        duration={{ enter: 'extra-long4', exit: 'short1' }}
         {...args}
       />
     </Stack>
-  )
-};
-
-export const TransitionOrigin: Story = {
-  render: (args) => (
-    <ScaleTemplate style={{ transformOrigin: 'top left' }} {...args} />
   )
 };
