+++ src/components/general/Transitions/Fade/index.ts
@@ -1,0 +1,1 @@
+export { default as Fade } from './Fade';
