+++ src/components/general/Transitions/Scale/index.ts
@@ -1,0 +1,1 @@
+export { default as Scale } from './Scale';
