+++ src/components/general/Transitions/Slide/Slide.mdx
@@ -1,0 +1,43 @@
+import { Meta, Canvas, ArgTypes, Stories } from '@storybook/blocks';
+import * as SlideStories from './Slide.stories';
+
+<Meta of={SlideStories} />
+
+# Slide
+
+Slide-In/Slide-Out Transition<br />
+(Transition 컴포넌트를 기반으로 만들어짐)
+
+## Import
+
+```js
+import { Slide } from '@/components/general/Transitions';
+```
+
+## Usage
+
+### Default Slide Transition
+
+- 화면의 edge에서부터 slide in 되고, 해당 edge로 slide out 됨
+
+<Canvas of={SlideStories.DefaultSlideTransition} />
+
+### Direction
+
+- `direction`: transition 방향
+  - 'down': 위 → 아래로 내려옴
+  - 'up': 아래 → 위로 올라옴
+  - 'left': 오른쪽 → 왼쪽으로 이동함
+  - 'right': 왼쪽 → 오른쪽으로 이동함
+
+<Canvas of={SlideStories.Direction} />
+
+### Slide Container
+
+- `container`: slide transition이 나타나는 container
+
+<Canvas of={SlideStories.SlideContainer} />
+
+## Props
+
+<ArgTypes of={SlideStories} sort="alpha" />
