+++ src/components/general/Transitions/Slide/index.ts
@@ -1,0 +1,1 @@
+export { default as Slide } from './Slide';
