+++ src/components/general/Transitions/index.ts
@@ -1,6 +1,6 @@
 export { default as Transition } from './Transition';
 export type { TransitionProps } from './Transition';
-export { default as Fade } from './Fade';
-export { default as Collapse } from './Collapse';
-export { default as Slide } from './Slide';
-export { default as Scale } from './Scale';
+export { Fade } from './Fade';
+export { Collapse } from './Collapse';
+export { Slide } from './Slide';
+export { Scale } from './Scale';
